[front/components/agent_builder/settings] - refactor: encapsulate agent settings sections into components
 - Extract agent name input logic into a dedicated `AgentBuilderNameSection` component
 - Create `AgentBuilderDescriptionSection` to handle agent description input and automatic suggestion
 - Implement `AgentBuilderAvatarSection` for selecting or auto-generating agent avatars
 - Simplify `AgentBuilderSettingsBlock` by utilizing new componentized sections for cleaner structure
